backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Liat <liathirsh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
flask = "^2.2.2"
sqlalchemy = "^1.4.46"
python-dotenv = "^0.21.1"
alembic = "^1.9.2"
flask-migrate = "^4.0.2"
gunicorn = "^20.1.0"
blueprint = "^3.4.2"
flask-cors = "^3.0.10"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
